Update volume control when gstreamer volume changes
Adjusting Pithos' volume through pulseaudio would change the the volume
in self.player, but not the volume control, resulting in the control
becoming out-of-sync with actual volume. This commit fixes this.

This should also fix https://bugs.launchpad.net/pithos/+bug/780031
since the volume preference will be updated for pulseaudio volume
changes.