Don't trigger the global play/pause on spacebar if the current focus is a button.
Previously pressing spacebar would cause a play/pause as well as whatever button behavior of the focused button. Worst case was when the play/pause button was focused, this would cause it to immediately toggle back to the previous state. Bottom line, when a button is focused, we should preserve its accessibility.