Handle nested signals in screensaver_pause.
In Unity, (and I would assume also in other combinations that will be
added in the future,) the lock signal and the screensaver signals occur
independently but can also occur at the same time. This keeps a count of
lock/unlock (screensaver on/off) pairs and ensures they are balanced
before resuming playing.