Better UI loop handling
As it sits Pithos starts a new UI loop every time play is pressed(maybe even more often?) and never destroys old loops. They stack up over time creating a small memory leak of sorts. This patch makes Pithos handle the UI loop more intelligently by starting/creating it when needed and stopping/destroying it when it's not needed (like when Pithos is paused). Credit for the code goes to the Exaile devs as it's pretty much a copy and paste of their code with a little tweaking to make it work with Pithos.