Tame the UI updates during buffering
AAC's spam the buffer messages with 0% and duplicate messages causing upwards of 1000+ buffer messages in less than 2 secs in some cases. This will calm it down a little, and make the buffer message in the UI a little more useful as it won't appear to be stuck on 0 most of the time.