Make use of the stream-start gstreamer message and clean up unneeded ones.
The songs duration can be got at and we can check for ads at the posting of the stream-start message. This allows us to remove async-done and duration-change. We can also git rid of the tag message(that spams the heck out of debug mode). The only thing we use it for is getting the bitrate which Pandora already gives us.