LastFM UI changes
Mostly UI changes with a couple small logic tweaks.

UI changes:
Created a unified look that matches the main prefs window, including "Reset" and "Close" buttons in the headerbar.

Logic Tweaks:
Added a "lastfm-authorized" signal to signal the plugin logic so it can act accordingly.
The plugin now updates LastFM with the current playing song if it's enabled in the middle of a song.
Got rid of the global for the worker. It wasn't needed, the worker can just be passed to the prefs dialog class.