mpris: Send empty return value for DBUS methods.
When something calls a DBUS method, it waits for a reply even for a method
that doesn't have a return value. So, if you don't send a (possibly empty)
return value, then the caller will hang until it times out.

So, in DBusServiceObject.__dbus_method_call, if the method doesn't have a
return value, then call g_dbus_method_invocation_return_value anyway, with
None for a parameter.