Changed PLAYLIST_VALIDITY_TIME in pandora.py to 1 hour.
While doing some testing for #351 I have determined that a playlist is only good for about an hour.
This also changes the is_still_valid method in pandora.py to account for the length of the song to determine if the song is still valid.
A song is considered valid if there is enough time for the song to play in it's entirety before the playlist has expired.