Check the validity of a song when going from pause to play.
Check the validity of a song when going from pause to play to more gracefully indicate that a song is no longer valid if it has become invalid while it was paused. Mostly useful in conjunction with the screensaver pause plugin or after other long periods in which Pithos might be left in a paused state. So when the user unlocked the screen Pithos would skip to a valid song without throwing a bunch of Gstreamer errors.