Find Missing Covers Plugin.
This Plugin attempts to find missing album cover art from Last.fm. If Album art can not be found an artist image is used instead if found.

Features:
Does NOT depend on pylast.
A UI for viewing, editing and optionally blacklisting artist and album corrections.
The corrections are searchable, sort-able and editable.
Art urls are cached and when reused are verified to still be valid.
Context menu to copy contents of cell and search Last.fm.