Improved buffering behavior
This improves buffering behavior in a few ways.

1st, It changes how we deal with buffering signals. Since the buffering percent that is received in the buffering message is pretty much useless this uses a buffering query to get a bool representing if the pipeline is buffering or not. We then respond to that bool accordingly.

2nd, It Enumerates the player states to give us a BUFFERING state. We can then use those Enums to clearly differentiate between PAUSED and BUFFERING.

3rd, It makes the timeout longer in souphttpsrc to give the user time to reconnect in the case of a network disconnect.

And finally it makes Pithos behave as you would expect during buffering as far as user initiated state changes. As in the pipeline doesn't actually change state until buffering completes at which time it goes to the correct state, whatever that may be.

Another positive side effect of this is that it puts the end to the UI thrashing during buffering.