Stops Pithos from quitting when loggin and Pandora One is changed
A bit of an edge case but I have 2 accounts one account is my main listening account, it is a Pandora Plus(or One) account the other is a free account. They both have the same password. If I change both the Pandora One checkbox and the loggin without changing the password Pithos quits with a "No email or no password set!" error. This stops that from happening and Pithos behaves as expected and logs into the other account.