Show dialog when trying to add station with existing seed.
As mentioned in https://github.com/pithos/pithos/issues/385 Pandora does not allow multiple stations with the same seed. If you try to create a station with an existing seed it will fail and Pandora will just return the station containing the exsisting seed. currently we do not check for this and we blindly just add the station to the station model, which if it already exsists leads to a duplicate station. This fixes that by checking if the returned station already exsists and if it does it pops up a dialog telling the user that a new station could not be created and gives them the option to listen to the exsisting station that contains the seed they were trying to use to create a new station with.