Refactor Last.fm plugin
Use a headerbar.

Better logging.

The auth dialog now pops up when the plugin is enabled but Pithos is not authorized with Last.fm.
The plugin is also disabled if the dialog is closed and Pithos is not authorized with Last.fm.
This is to avoid confusion as to if the plugin is actually really enabled.

Use pylast.LastFMNetwork preferably instead of pylast.get_lastfm_network.