Offer to disable and deauthorize Last.fm plugin on login change and deauthorize on disable
Show a dialog with a link to the Last.fm account that the plugin is authorized for and lets them decide if they want to continue to use that account on a Pandora login change if the plugin is enabled else it else if the plugin is disabled it gives them the option to deauthorize if it is authorized. It also shows the same deauthorize dialog on disabling the plugin.

Fixes: https://github.com/pithos/pithos/issues/442