Handle not being able to rename Thumbprint Radio
Pandora does not permit users to rename the Thumbprint Radio Station. This handles that by catching the appropriate Pandora error and showing a dialog explaining that to the user and resetting the name back to the old name in the stations dialog. Fixes: https://github.com/pithos/pithos/issues/370 for the most part.