Detect the Thumbprint Radio Station
This will make it so we treat the Thumbprint Radio similar to the QuickMix. It will always be right after the QuickMix in the stations popover and will not show up in the station dialog since it can't be renamed. It will also always be the 2nd station in MPRIS.