First steps in making plugins GObjects
This is the 1st step in making plugins GObjects. This allows us to potentially to do a bunch of neat stuff. For now this particular commit fixes a bug where when enabling a plugin with a prefs window the prefs button is not sensitive until you click off somewhere else and back on that plugin. This was caused because our plugins now load async.