Don't GoTo songs that are banned or tired.
A user shouldn't be able to GoTo a song that is banned or tired because Pithos will just skip that banned or tired song and play the next song. In the case the song is banned or tired it makes more sense to do nothing. It's more of an expected behaviour then playing a song that wasn't selected(the next song).