song.title should not be a property
If song.title is a decorated property with a getter that means that the 1st thing to access it has to potentially do a blocking network call. It just so happens that the 1st thing to access song.title does so from the main thread. Doing blocking calls from the main thread is ofc bad. This instead moves checking the title to init to make sure that any blocking calls are done from a worker thread.