We can't use the GstAudio constants
As much as I would prefer to use the GstAudio.AUDIO_RESAMPLER_QUALITY_MAX and GstAudio.AudioResamplerFilterMode.FULL constants, they don't exist in the version of Gstreamer that ships with *buntu 16.04. Trying to use them in *buntu 16.04 crashes Pithos.

Instead just set "quality" and "sinc-filter-mode" to the values the constants would be if they existed...