Raise the priority of the GLib.idle_add's in the worker to GLib.PRIORITY_DEFAULT
As it sits now the entry priority of worker threads is really low. Worker threads have to wait for time to refresh in the UI for example because it's a higher priority.

This commit has the effect of making things that use the worker thread generally more "snappy". Songs load faster in the UI, station searches are faster, song ratings are faster.